+++ android/app/src/main/java/com/brekeke/phonedev/BrekekeUtils.java
@@ -120,6 +120,7 @@ public class BrekekeUtils extends ReactContextBaseJavaModule {
   public static boolean isAppActiveLocked = false;
   public static boolean firstShowCallAppActive = false;
   public static boolean phoneappliEnabled = false;
+  public static String userAgentConfig = null;
 
   BrekekeUtils(ReactApplicationContext c) {
     super(c);
@@ -722,7 +723,7 @@ public class BrekekeUtils extends ReactContextBaseJavaModule {
     ctx.startActivity(intent);
   }
 
-  // Overlay screen permission
+  // overlay screen permission
   public static boolean isOverlayPermissionGranted(Context context) {
     if (Build.VERSION.SDK_INT < Build.VERSION_CODES.M) {
       return true;
@@ -746,8 +747,38 @@ public class BrekekeUtils extends ReactContextBaseJavaModule {
     }
   }
 
+  public static boolean isUserAgentConfig() {
+    return BrekekeUtils.userAgentConfig != null && !BrekekeUtils.userAgentConfig.equals("");
+  }
+
   // ==========================================================================
   // react methods
+
+  @ReactMethod
+  public void setUserAgentConfig(String userAgentConfig) {
+    if (BrekekeUtils.userAgentConfig == null) {
+      UiThreadUtil.runOnUiThread(
+          new Runnable() {
+            @Override
+            public void run() {
+              try {
+                for (IncomingCallActivity a : activities) {
+                  try {
+                    a.updateUserAgentConfig(userAgentConfig);
+                  } catch (Exception e) {
+                    BrekekeUtils.emit(
+                        "debug", "IncomingCallActivity::updateUserAgentConfig " + e.getMessage());
+                  }
+                }
+              } catch (Exception e) {
+                BrekekeUtils.emit(
+                    "debug", "IncomingCallActivity::updateUserAgentConfig " + e.getMessage());
+              }
+            }
+          });
+    }
+    BrekekeUtils.userAgentConfig = userAgentConfig;
+  }
 
   @ReactMethod
   public void isOverlayPermissionGranted(Promise p) {