+++ src/polyfill/index.ts
@@ -1,6 +1,7 @@
 import 'react-native-get-random-values'
 import './shared'
 
+import BgTimer from 'react-native-background-timer'
 import {
   mediaDevices,
   MediaStream,
@@ -38,7 +39,9 @@ window.navigator.getUserMedia =
   // @ts-ignore
   window.navigator.getUserMedia || mediaDevices.getUserMedia
 
-if (!window.BackgroundTimer) {
-  // type-coverage:ignore-next-line
-  window.BackgroundTimer = BackgroundTimer
-}
+// @ts-ignore
+window.setTimeout = (f, t) => BgTimer.setTimeout(f, Math.floor(t) || 0)
+window.clearTimeout = id => BgTimer.clearTimeout(id)
+// @ts-ignore
+window.setInterval = (f, t) => BgTimer.setInterval(f, Math.floor(t) || 0)
+window.clearInterval = id => BgTimer.clearInterval(id)