+++ src/stores/Call.ts
@@ -1,14 +1,19 @@
-import { action, observable } from 'mobx'
+import CiruclarJSON from 'circular-json'
+import type { IReactionDisposer } from 'mobx'
+import { action, autorun, observable } from 'mobx'
 import { Platform } from 'react-native'
 import RNCallKeep from 'react-native-callkeep'
 
 import { pbx } from '../api/pbx'
 import { sip } from '../api/sip'
 import type { Session, SessionStatus } from '../brekekejs'
+import { embedApi } from '../embed/embedApi'
 import { getPartyName } from '../stores/contactStore'
 import { checkPermForCall } from '../utils/permissions'
 import { BrekekeUtils } from '../utils/RnNativeModules'
 import { waitTimeout } from '../utils/waitTimeout'
+import { authPBX } from './AuthPBX'
+import { getAuthStore } from './authStore'
 import type { CallStore } from './callStore2'
 import { contactStore } from './contactStore'
 import { intlDebug } from './intl'
@@ -17,7 +22,7 @@ import { RnAlert } from './RnAlert'
 
 export class Call {
   constructor(private store: CallStore) {}
-
+  line?: string
   rawSession?: Session
 
   @observable earlyMedia: MediaStream | null = null
@@ -61,7 +66,7 @@ export class Call {
   answer = async (
     options?: { ignoreNav?: boolean },
     videoOptions?: object,
-    exInfo?: object,
+    exInfo?: string,
   ) => {
     this.holding = false
     this.answered = true
@@ -131,6 +136,7 @@ export class Call {
 
   // to use in embed api and hang up special transfer case
   hangup = () => {
+    this.isAboutToHangup = true
     sip.hangupSession(this.id)
   }
 
@@ -218,6 +224,8 @@ export class Call {
         : intlDebug`Failed to start recording the call`
       RnAlert.error({ message, err })
     }
+    // try to re-auth if error is timeout
+    this.checkTimeoutToReconnectPbx(err)
   }
 
   toggleHoldWithCheck = () => {
@@ -236,14 +244,34 @@ export class Call {
     if (!this.isAboutToHangup && fn === 'unhold') {
       this.store.setCurrentCallId(this.id)
     }
+
     return pbx[`${fn}Talker`](this.pbxTenant, this.pbxTalkerId)
       .then(this.onToggleHoldFailure)
       .catch(this.onToggleHoldFailure)
+  }
+
+  private checkTimeoutToReconnectPbx = (err: Error | boolean) => {
+    if (err === true) {
+      return
+    }
+    if (
+      err &&
+      typeof err === 'object' &&
+      (('code' in err && (err as any).code === -1) ||
+        ('message' in err && /timeout/i.test(err.message)))
+    ) {
+      getAuthStore().pbxState = 'stopped'
+      authPBX.dispose()
+      authPBX.auth()
+    }
   }
   @action private onToggleHoldFailure = (err: Error | boolean) => {
     if (err === true) {
       return true
     }
+    // try to re-auth if error is timeout
+    this.checkTimeoutToReconnectPbx(err)
+
     const prevFn = this.holding ? 'hold' : 'unhold'
     this.setHolding(prevFn === 'unhold')
     if (typeof err !== 'boolean') {
@@ -292,6 +320,8 @@ export class Call {
       message: intlDebug`Failed to transfer the call`,
       err,
     })
+    // try to re-auth if error is timeout
+    this.checkTimeoutToReconnectPbx(err)
   }
 
   @action stopTransferring = () => {
@@ -310,6 +340,8 @@ export class Call {
       message: intlDebug`Failed to stop the transfer`,
       err,
     })
+    // try to re-auth if error is timeout
+    this.checkTimeoutToReconnectPbx(err)
   }
 
   @action conferenceTransferring = () => {
@@ -328,6 +360,8 @@ export class Call {
       message: intlDebug`Failed to make conference for the transfer`,
       err,
     })
+    // try to re-auth if error is timeout
+    this.checkTimeoutToReconnectPbx(err)
   }
 
   @action park = (number: string) =>
@@ -339,6 +373,41 @@ export class Call {
       message: intlDebug`Failed to park the call`,
       err,
     })
+    // try to re-auth if error is timeout
+    this.checkTimeoutToReconnectPbx(err)
+  }
+
+  private _autorunEmitEmbed = false // check if autorun is already started
+  private _disposeEmitEmbed?: IReactionDisposer // dispose autorun
+  startEmitEmbed = () => {
+    if (window._BrekekePhoneWebRoot) {
+      return
+    }
+    embedApi.emit('call', this)
+    this._disposeEmitEmbed = autorun(() => {
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
+      const { store, ...c } = this // do not autorun on store
+      CiruclarJSON.stringify(c)
+      if (!this._autorunEmitEmbed) {
+        this._autorunEmitEmbed = true
+        return
+      }
+      if (this.isAboutToHangup) {
+        return
+      }
+      embedApi.emit('call_update', this)
+    })
+  }
+  disposeEmitEmbed = () => {
+    this._disposeEmitEmbed?.()
+    this._disposeEmitEmbed = undefined
+  }
+  finishEmitEmbed = () => {
+    if (window._BrekekePhoneWebRoot) {
+      return
+    }
+    this.disposeEmitEmbed()
+    embedApi.emit('call_end', this)
   }
 }