+++ src/stores/authStore2.ts
@@ -2,10 +2,11 @@ import { debounce } from 'lodash'
 import { action, observable } from 'mobx'
 import { AppState, Platform } from 'react-native'
 
-import { sip } from '../api/sip'
+import { getUserAgent, sip } from '../api/sip'
 import type {
   PbxCustomPage,
   PbxGetProductInfoRes,
+  PbxResourceLine,
   UcBuddy,
   UcBuddyGroup,
   UcConfig,
@@ -14,9 +15,10 @@ import { BackgroundTimer } from '../utils/BackgroundTimer'
 import { clearUrlParams, getUrlParams } from '../utils/deeplink'
 import type { ParsedPn, SipPn } from '../utils/PushNotification-parse'
 import { BrekekeUtils } from '../utils/RnNativeModules'
+import { toBoolean } from '../utils/string'
 import { waitForActiveAppState } from '../utils/waitForActiveAppState'
 import { waitTimeout } from '../utils/waitTimeout'
-import type { Account } from './accountStore'
+import type { Account, RecentCall } from './accountStore'
 import {
   accountStore,
   getAccountUniqueId,
@@ -57,10 +59,16 @@ export class AuthStore {
   @observable sipTotalFailure = 0
   @observable ucState: ConnectionState = 'stopped'
   @observable ucTotalFailure = 0
+
+  @observable pbxLoginFromAnotherPlace = false
+  @observable showMsgPbxLoginFromAnotherPlace = false
   @observable ucLoginFromAnotherPlace = false
+
+  @observable pbxConnectedAt = 0
 
   pbxShouldAuth = () =>
     this.getCurrentAccount() &&
+    !this.pbxLoginFromAnotherPlace &&
     this.pbxState !== 'waiting' &&
     // do not auth pbx if sip token is provided in case of PN
     // wait until sip login success or failure
@@ -70,12 +78,17 @@ export class AuthStore {
     (this.pbxState === 'stopped' ||
       (this.pbxState === 'failure' &&
         // !this.pbxTotalFailure &&
-        RnAppState.currentState === 'active'))
+        RnAppState.currentState === 'active') ||
+      (this.pbxState === 'failure' &&
+        this.sipState === 'success' &&
+        RnAppState.currentState === 'background' &&
+        getCallStore().calls.length))
   pbxConnectingOrFailure = () =>
     ['waiting', 'connecting', 'failure'].some(s => s === this.pbxState)
 
   sipShouldAuth = () =>
     this.sipState !== 'waiting' &&
+    !this.pbxLoginFromAnotherPlace &&
     this.sipState !== 'connecting' &&
     this.sipState !== 'success' &&
     ((this.signedInId && this.sipPn.sipAuth) ||
@@ -101,12 +114,17 @@ export class AuthStore {
     this.getCurrentAccount()?.ucEnabled &&
     ['waiting', 'connecting', 'failure'].some(s => s === this.ucState)
 
-  isConnFailure = () =>
-    [
+  isConnFailure = (): boolean => {
+    if (this.pbxLoginFromAnotherPlace || this.ucLoginFromAnotherPlace) {
+      return true
+    }
+    const states = [
       this.pbxState,
       this.sipState,
-      this.getCurrentAccount()?.ucEnabled && this.ucState,
-    ].some(s => s === 'failure')
+      this.getCurrentAccount()?.ucEnabled ? this.ucState : undefined,
+    ].filter(s => !!s)
+    return !states.includes('connecting') && states.includes('failure')
+  }
 
   @observable signedInId = ''
   getCurrentAccount = () =>
@@ -120,7 +138,6 @@ export class AuthStore {
   @observable ucConfig?: UcConfig
   @observable pbxConfig?: PbxGetProductInfoRes
   @observable listCustomPage: PbxCustomPage[] = []
-
   saveActionOpenCustomPage = false
   customPageLoadings: { [k: string]: boolean } = {}
   getCustomPageById = (id: string) => this.listCustomPage.find(i => i.id == id)
@@ -131,6 +148,26 @@ export class AuthStore {
     }
     Object.assign(found, cp)
   }
+
+  @observable resourceLines: PbxResourceLine[] = []
+
+  @observable userAgentConfig: string | undefined = undefined
+
+  setUserAgentConfig = async (useragent: string | undefined) => {
+    const isEnabled = useragent === undefined || toBoolean(useragent)
+    if (!isEnabled) {
+      BrekekeUtils.setUserAgentConfig('')
+      return
+    }
+    const a = this.getCurrentAccount()
+    if (!a) {
+      return
+    }
+    const userAgent = await getUserAgent(a)
+    BrekekeUtils.setUserAgentConfig(userAgent)
+    this.userAgentConfig = userAgent
+  }
+
   isBigMode = () => this.pbxConfig?.['webphone.allusers'] === 'false'
 
   signIn = async (a?: Account, autoSignIn?: boolean) => {
@@ -167,12 +204,14 @@ export class AuthStore {
     }
 
     this.signedInId = a.id
+    this.pbxConnectedAt = 0
     BrekekeUtils.setPhoneappliEnabled(!!this.phoneappliEnabled())
     if (!autoSignIn) {
       await saveLastSignedInId(getAccountUniqueId(a))
     }
     return true
   }
+
   autoSignInLast = async () => {
     const d = await getLastSignedInId()
     const a = await accountStore.findByUniqueId(d.id)
@@ -225,7 +264,7 @@ export class AuthStore {
     this.sipPn = {}
     sip.stopWebRTC()
     this.ucState = 'stopped'
-    this.resetFailureStateIncludeUcLoginFromAnotherPlace()
+    this.resetFailureStateIncludePbxOrUc()
     this.pbxConfig = undefined
     this.ucConfig = undefined
     this.listCustomPage = []
@@ -234,6 +273,9 @@ export class AuthStore {
     contactStore.clearStore()
     chatStore.clearStore()
     this.userExtensionProperties = null
+    this.cRecentCalls = []
+    this.rcPage = 0
+    this.pbxConnectedAt = 0
   }
 
   @action resetFailureState = () => {
@@ -259,21 +301,113 @@ export class AuthStore {
     this.sipState = 'stopped'
     authSIP.auth()
   }
-  @action resetFailureStateIncludeUcLoginFromAnotherPlace = () => {
+
+  @action resetFailureStateIncludePbxOrUc = () => {
     this.resetFailureState()
-    this.ucLoginFromAnotherPlace = false
-    authUC.auth()
+    if (this.pbxLoginFromAnotherPlace) {
+      this.pbxLoginFromAnotherPlace = false
+      this.showMsgPbxLoginFromAnotherPlace = false
+      authPBX.auth()
+      getAuthStore().pbxConnectedAt = 0
+    }
+    if (this.ucLoginFromAnotherPlace) {
+      this.ucLoginFromAnotherPlace = false
+      authUC.auth()
+    }
   }
 
+  recentCallsMax: number = 200
+  @observable cRecentCalls: RecentCall[] = []
+  rcPerPage: number = 15
+  rcPage: number = 0
+  @observable rcLoading: boolean = false
+  @observable rcCount: number = 0
+  // recentCallsMax with default 200 and limit 1000
+  setRecentCallsMax = async (max: number | string) => {
+    const numericMax = Number(max)
+
+    this.recentCallsMax =
+      Number.isInteger(numericMax) && numericMax > 0
+        ? Math.min(numericMax, 1000)
+        : 200
+
+    // update recentCalls if config recentCallsMax changed
+    const d = await this.getCurrentDataAsync()
+    if (!d) {
+      return
+    }
+    if (d.recentCalls.length > this.recentCallsMax) {
+      d.recentCalls.splice(this.recentCallsMax)
+      this.rcCount = d.recentCalls.length
+      accountStore.saveAccountsToLocalStorageDebounced()
+    }
+  }
+
+  rcFirstTimeLoadData = async () => {
+    if (this.rcPage === 0) {
+      this.rcLoading = true
+      const d = this.getCurrentData()
+      if (!d) {
+        return
+      }
+      const filteredCalls =
+        d?.recentCalls.filter(this.isMatchUserRecentCalls) || []
+      const calls = filteredCalls.slice(0, this.rcPerPage) || []
+      this.cRecentCalls = calls
+      this.rcCount = filteredCalls.length
+      this.rcLoading = false
+    }
+  }
+  isMatchUserRecentCalls = (call: RecentCall) => {
+    if (call.partyNumber.includes(contactStore.callSearchRecents)) {
+      return call.id
+    }
+    return ''
+  }
+  rcSearchRecentCall = async () => {
+    const d = this.getCurrentData()
+    if (!d) {
+      return
+    }
+    this.rcPage = 0
+    this.rcLoading = true
+    this.cRecentCalls = []
+    const filteredCalls =
+      d?.recentCalls.filter(this.isMatchUserRecentCalls) || []
+    this.cRecentCalls = filteredCalls.slice(0, this.rcPerPage)
+    this.rcCount = filteredCalls.length
+    this.rcLoading = false
+  }
+  rcLoadMore = () => {
+    this.rcLoading = true
+    const d = this.getCurrentData()
+    if (!d) {
+      return
+    }
+    this.rcPage++
+    const calls =
+      d?.recentCalls
+        .filter(this.isMatchUserRecentCalls)
+        .slice(
+          this.rcPerPage * this.rcPage,
+          this.rcPerPage * (this.rcPage + 1),
+        ) || []
+    this.cRecentCalls = [...this.cRecentCalls, ...calls]
+    this.rcLoading = false
+  }
   pushRecentCall = async (call: CallHistoryInfo) => {
     const d = await this.getCurrentDataAsync()
     if (!d) {
       return
     }
+
     d.recentCalls = [call, ...d.recentCalls]
-    if (d.recentCalls.length > 20) {
-      d.recentCalls.pop()
+    this.cRecentCalls = [call, ...this.cRecentCalls]
+    if (d.recentCalls.length > this.recentCallsMax) {
+      d.recentCalls.splice(this.recentCallsMax)
+      this.cRecentCalls.splice(this.recentCallsMax)
     }
+    this.rcCount = d.recentCalls.length
     accountStore.saveAccountsToLocalStorageDebounced()
   }